+++ packages/ui/src/components/sign-in/index.tsx
@@ -1,1 +1,1 @@
-export { SignInComponent as SignIn } from './sign-in';
+export * from './sign-in';
