+++ packages/ui/src/components/sign-up/index.ts
@@ -1,0 +1,1 @@
+export * from './sign-up';
